Adaption to timm 0.9.x naming scheme
In 'timm 0.9.x', the model naming scheme has changed, e.g. 'vit_base_patch16_224_in21k' -> 'vit_base_patch16_224.augreg_in21k' (the datasets are now separated by a '.'), but the part after '.' is omitted in 'timm.list_models()'. 
The change of the if-statement is a minimal change that allows for compatibility to both versions of 'timm'.